Add slow fall effect to cloud armor
Add a the new class CustomArmorItem which checks if the full set is on and gives the effect. This class has two abstract methods: getItemclass() and getArmorEffect() these are the methods which are individual for every armor. To use the CustomArmorItem class add a new class for each armor which extends the CustomArmorClass and override the abstract methods. In the getItemClass() return this.getClass() and in the getArmorEffec() return the effect you want the armor to have. To avoid maintaning the wrong amplifier when the player wears full set while giving itself slow fall with a higher amplifier, the CustomArmorItem class holds a if/else statement to check for this case and then giving a new EffectInstance.